New HR Single Verifier
